create auth provider

+++ src/db/mutations/useDeleteBook.ts
@@ -1,10 +1,12 @@
-import { supabase } from "..";
 import { BOOK_TABLE } from "../schemas/book";
 import { AbstractPowerSyncDatabase, usePowerSync } from "@powersync/react-native";
 import { useState } from "react";
 
-const deleteBook = async (bookId: string, userId: string | null, db: AbstractPowerSyncDatabase) => {
-  await db.execute(`DELETE FROM ${BOOK_TABLE} WHERE id = ? AND user_id = ?`, [bookId, userId]);
+import { LOCAL_USER_ID } from "@/consts";
+import { useAuth } from "@/providers/AuthProvider";
+
+const deleteBook = async (bookId: string, db: AbstractPowerSyncDatabase, userId?: string) => {
+  await db.execute(`DELETE FROM ${BOOK_TABLE} WHERE id = ? AND user_id = ?`, [bookId, userId || LOCAL_USER_ID]);
 
   // const resultRecord = res.rows?.item(0);
   // if (!resultRecord) {
@@ -15,13 +17,13 @@ const deleteBook = async (bookId: string, userId: string | null, db: AbstractPow
 
 export function useDeleteBook() {
   const [isMutating, setIsMutating] = useState(false);
+  const { user } = useAuth();
 
   const db = usePowerSync();
 
   async function mutate(mealId: string) {
     setIsMutating(true);
-    const userId = await supabase.getUserId();
-    const result = await deleteBook(mealId, userId, db);
+    const result = await deleteBook(mealId, db, user?.id);
     setIsMutating(false);
     return result;
   }
