+++ src/db/mutations/useDeleteMeal.ts
@@ -1,10 +1,12 @@
-import { supabase } from "..";
 import { MEAL_TABLE } from "../schemas/meal";
 import { AbstractPowerSyncDatabase, usePowerSync } from "@powersync/react-native";
 import { useState } from "react";
 
-const deleteMeal = async (mealId: string, userId: string | null, db: AbstractPowerSyncDatabase) => {
-  await db.execute(`DELETE FROM ${MEAL_TABLE} WHERE id = ? AND user_id = ?`, [mealId, userId]);
+import { LOCAL_USER_ID } from "@/consts";
+import { useAuth } from "@/providers/AuthProvider";
+
+const deleteMeal = async (mealId: string, db: AbstractPowerSyncDatabase, userId?: string) => {
+  await db.execute(`DELETE FROM ${MEAL_TABLE} WHERE id = ? AND user_id = ?`, [mealId, userId || LOCAL_USER_ID]);
 
   // const resultRecord = res.rows?.item(0);
   // if (!resultRecord) {
@@ -15,13 +17,13 @@ const deleteMeal = async (mealId: string, userId: string | null, db: AbstractPow
 
 export function useDeleteMeal() {
   const [isMutating, setIsMutating] = useState(false);
+  const { user } = useAuth();
 
   const db = usePowerSync();
 
   async function mutate(mealId: string) {
     setIsMutating(true);
-    const userId = await supabase.getUserId();
-    const result = await deleteMeal(mealId, userId, db);
+    const result = await deleteMeal(mealId, db, user?.id);
     setIsMutating(false);
     return result;
   }
